Don't indent invisible template slices
Some custom templater plugins might emit special zero-length marker
template slices with a special slice type, as previously described in
#4708.  These are transformed into a meta TemplateSegment with an empty
source string, which mostly works great.  However, the LT02 indention
rule will attempt to indent these meta nodes, which is impossible
because there is no source_str that the indention would affect, and thus
it is impossible to resolve this lint violation.

For example, suppose the parse tree looks like this:

[META] placeholder: [Type: 'block_start', Raw: "{% up 'Do nothing' %}", Block: 'f142fe']
[META] indent:      [Block: 'f142fe']
[META] placeholder: [Type: 'migration_up', Raw: '']
newline:            '\n'
whitespace:         '    '

The indention rule will perceive a violation and try to fix it by
inserting a newline and whitespace prior to the second placeholder:

[META] placeholder: [Type: 'block_start', Raw: "{% up 'Do nothing' %}", Block: 'f142fe']
[META] indent:      [Block: 'f142fe']
newline:            '\n'
whitespace:         '    '
[META] placeholder: [Type: 'migration_up', Raw: '']
newline:            '\n'
whitespace:         '    '

However, this second placeholder doesn't have any source string.  When
the file is reparsed, the templater will again result in a parse tree
as seen in the original snippet above, and thus the violation is
effectively unfixable.

This commit fixes the issue by treating these cases identically to
trailing comments at the end of a line.  For example, the following
template has a trailing comment that delays the indention requirement:

{% if True %} -- I don't need to put a trailing comment on the next line
    SELECT 1;

By treating the zero-length marker placeholder the same as a trailing
comment, we can avoid trying to indent them.

Note that this commit also adds test cases to the reindent_test for
trailing comments.  Previously, there was no direct test coverage for
that branch of code.